+++ backend/hqlib/metric/document/age.py
@@ -34,7 +34,7 @@ class DocumentAge(LowerIsBetterMetric):
     missing_template = 'Het document "{name}" is niet aangetroffen.'
     target_value = 180
     low_target_value = 200
-    metric_source_class = metric_source.ArchiveSystem
+    metric_source_class = metric_source.VersionControlSystem
 
     def value(self):
         """ Return the number of days since the document was last changed. """